Fix pano-kill scope

+++ home/lg/earth/scripts/pano-kill.sh
@@ -2,6 +2,8 @@
 
 . ${HOME}/etc/shell.conf
 
-lg-run "killall -9 pano-launcher.sh"
-lg-run-bg "killall -9 xiv"
-lg-sudo-bg "killall -CONT googleearth-bin"
+if [[ "${FRAME_NO}" == "0" ]]; then
+    lg-run "killall -9 pano-launcher.sh"
+    lg-run-bg "killall -9 xiv"
+    lg-sudo-bg "killall -CONT googleearth-bin"
+fi
